Crate todo_list

Structs

App 🔒

This struct holds the current state of the app. In particular, it has the todo_list field which is a wrapper around ListState. Keeping track of the state lets us render the associated widget with its state and have access to features such as natural scrolling.

TodoItem 🔒

TodoList 🔒

Enums

Status 🔒

Constants

ALT_ROW_BG_COLOR 🔒

COMPLETED_TEXT_FG_COLOR 🔒

NORMAL_ROW_BG 🔒

SELECTED_STYLE 🔒

TEXT_FG_COLOR 🔒

TODO_HEADER_STYLE 🔒

Functions

alternate_colors 🔒

main 🔒